backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
openai==1.3.7
anthropic==0.7.7
sentence-transformers==2.2.2
qdrant-client==1.7.0
numpy==1.24.3
tiktoken==0.5.2
aiofiles==23.2.0
websockets==12.0
redis==5.0.1
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
jinja2==3.1.2
gitpython==3.1.40
tree-sitter==0.20.4
tree-sitter-python==0.20.4
tree-sitter-javascript==0.20.1
tree-sitter-typescript==0.20.2
tree-sitter-java==0.20.2
tree-sitter-cpp==0.20.0
tree-sitter-go==0.20.0
tree-sitter-rust==0.20.4
